requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
flask==2.3.3
requests==2.31.0
psutil==5.9.5
matplotlib==3.7.2
pandas==2.0.3
line-bot-sdk==3.5.0
Pillow==10.0.0
